stripping more oo leftovers
