--- object/comparable.go
@@ -1,9 +1,0 @@
-package object
-
-var comparableModule = newModule("Comparable", comparableMethodSet, nil)
-
-func init() {
-	classes.Set("Comparable", comparableModule)
-}
-
-var comparableMethodSet = map[string]RubyMethod{}
