+++ object/introspect.go
@@ -1,13 +1,11 @@
 package object
 
-func getMethods(class RubyClass, visibility MethodVisibility, addSuperMethods bool) *Array {
+func getMethods(class RubyClass, addSuperMethods bool) *Array {
 	var methodSymbols []RubyObject
 	for class != nil {
 		methods := class.Methods().GetAll()
-		for meth, fn := range methods {
-			if fn.Visibility() == visibility {
-				methodSymbols = append(methodSymbols, &Symbol{meth})
-			}
+		for meth, _ := range methods {
+			methodSymbols = append(methodSymbols, &Symbol{meth})
 		}
 		if !addSuperMethods {
 			break
