+++ object/send.go
@@ -1,9 +1,5 @@
 package object
 
-import (
-	"github.com/pkg/errors"
-)
-
 // Send sends message method with args to context and returns its result
 func Send(context CallContext, method string, args ...RubyObject) (RubyObject, error) {
 	receiver := context.Receiver()
@@ -15,10 +11,6 @@ func Send(context CallContext, method string, args ...RubyObject) (RubyObject, e
 		if !ok {
 			class = class.SuperClass()
 			continue
-		}
-
-		if fn.Visibility() == PRIVATE_METHOD && receiver.Type() != SELF {
-			return nil, errors.WithStack(NewPrivateNoMethodError(receiver, method))
 		}
 
 		return fn.Call(context, args...)
